feat: integrate BackupsModal into ExcalidrawWrapper and MainMenu
- Added BackupsModal component to ExcalidrawWrapper for managing canvas backups.
- Updated MainMenuConfig to include handlers for showing and closing the BackupsModal.
- Enhanced styles for the AuthDialog and BackupsModal to improve UI consistency.
- Adjusted BackupsModal properties for better user interaction, including size and close behavior.
